+++ app/src/main/java/com/example/financepal/Balance.java
@@ -72,7 +72,7 @@ public class Balance extends AppCompatActivity {
 
 
         DbNombreMetas dbNombreMetas = new DbNombreMetas(this);
-        Double montomensualmetas = dbNombreMetas.sumarMontosMensuales(correoElectronicoS);
+        double montomensualmetas = dbNombreMetas.sumarMontosMensuales(correoElectronicoS);
         espacioMetasMensualesTotales.setText((col.format(montomensualmetas)) + " COP");
 
 
@@ -90,7 +90,7 @@ public class Balance extends AppCompatActivity {
         UsuarioGastos gastoAltoUsuario = (dbGastos.gastoMasAlto(correoElectronicoS));
         if (gastoAltoUsuario != null) {
             String nombreGastoMasAlto = gastoAltoUsuario.getNombregasto();
-            int cantGastoMasAlto = gastoAltoUsuario.getMontogasto();
+            long cantGastoMasAlto = gastoAltoUsuario.getMontogasto();
             espacioMayorGasto.setText(nombreGastoMasAlto + "\n" + (col.format(cantGastoMasAlto) + " COP"));
         } else {
             espacioMayorGasto.setText("No hay gastos relacionados");
@@ -100,7 +100,7 @@ public class Balance extends AppCompatActivity {
         UsuarioGastos gastoRecurrenteUsuario = dbGastos.gastoMasRecurrente(correoElectronicoS);
         if (gastoRecurrenteUsuario != null) {
             String nombreGastoMasRecurrente = gastoRecurrenteUsuario.getNombregasto();
-            int cantGastoMasREcurrente = gastoRecurrenteUsuario.getMontogasto()*gastoRecurrenteUsuario.getRecurrenciagasto();
+            long cantGastoMasREcurrente = gastoRecurrenteUsuario.getMontogasto()*gastoRecurrenteUsuario.getRecurrenciagasto();
             espacioGastoMasRecurrente.setText(nombreGastoMasRecurrente+"\n"+(col.format(cantGastoMasREcurrente) + " COP"));
         } else {
             espacioGastoMasRecurrente.setText("No hay gastos relacionados");
@@ -124,7 +124,7 @@ public class Balance extends AppCompatActivity {
         UsuarioGastos gastoPrioridad = dbGastos.gastoMasPrioridades(correoElectronicoS);
         if (gastoPrioridad != null) {
             String nombreGastoMasPrioridad = gastoPrioridad.getNombregasto();
-            int cantGastoMasPrioridad = gastoPrioridad.getMontogasto();
+            long cantGastoMasPrioridad = gastoPrioridad.getMontogasto();
             espacioGastoMenorPrioridad.setText(nombreGastoMasPrioridad+"\n"+(col.format(cantGastoMasPrioridad) + " COL"));
         } else {
             espacioGastoMenorPrioridad.setText("No hay gastos relacionados");
